+++ api_ai_translate_text.go
@@ -65,7 +65,7 @@ type TranslateTextReqGlossary struct {
 }
 
 type translateTextResp struct {
-	Code int                `json:"code,omitempty"` // 错误码，非 0 表示失败
+	Code int64              `json:"code,omitempty"` // 错误码，非 0 表示失败
 	Msg  string             `json:"msg,omitempty"`  // 错误描述
 	Data *TranslateTextResp `json:"data,omitempty"` //
 }
